Erlang 18->18.2; copy ansible into every image
Previously, the Dockerfile only copied the ansible/ directory into the
-base images. This means that if you want to add a new version of Erlang
(for instance), you have to rebuild *all* of the images just to populate
the /ansible directory to the child images.

So instead we simply re-ADD the /ansible directory into each child
image. This allows for generation of new os-erlang-XX.X images (or other
things, in the future) without respinning every single image.