Simplify Debian 11 support
When we first introduced support for Bullseye we added some logic to
handle Erlang 19 and below. We no longer support those obsolete
versions, and the logic we added is now actively harmful (e.g. the use
of gcc 9 instead of gcc 10 caused us to build Erlang 24 without the
much-anticipated JIT compiler).

This patch reverts back to defaults for Bullseye. It has the side effect
of causing Erlang 20 builds to fail to link properly, possibly because
of the switch to `-fno-common` by default in gcc 10. Given that Erlang
20 is already unsupported by the Erlang team, I didn't bother to try to
fix it here.