Restore old xplat support, add new buildx targets
This preserves the method of generating single-arch images while
adding targets that use the buildx plugin to generate multi-arch images.

In the future it would be possible to combine the single-arch images
into a multi-arch one using the manifest approach if someone were so
inclined. The couchdb-docker repo has the code to do this for the image
we provide to end users.

This commit is meant to revert many of the changes from 5c78da4 and can
be reviewed together with it.