Make sure Elixir compiled OTP version matches the current Erlang version
Currently Elixir artifacts are compiled on Erlang 22.  That seems to
work with Erlang/OTP 22 through 24 but doesn't work with Erlang 25
where it produces this error:

```
[error] beam/beam_load.c(86): Error loading module 'Elixir.Code.Formatter':
 lambda already defined for label 749. To fix this, please recompile this module with an OTP 25 compiler.
```

To fix it, downloading the Elixir version compiled with the same major
version of Erlang. This is what the asdf version manager does so we do
the same thing.