Update default Erlang and cleanup unused commands
Some of those haven't worked in more than 2 years and we've only used
buildx to build both the x86 and the multi-arch images.

Ideally we wouldn't want to use proprietary Docker bits but that's the
only one that works well with multi-arch builds. Perhaps Podman or
other newer things can replace it soon enough.