Update pull-all-couchdbdev-docker script
It turns out were were pruning all the active images as we a bunch of stale
Erlang 19 and 20 ones in the list to keep. Update the list so it's current and
add a note to the list OS image variables as a reminder to keep the cleanup
script in sync.

This should hopefully save 2-3 minutes on each CI run as the most of the docker
images should already be there. Previously it took about that much it looks
like to fetch the images on every CI run after the prune script ran.